+++ src/main/java/com/cafe/erp/modules/inventory/material/dto/MaterialUpdateRequest.java
@@ -3,6 +3,8 @@ package com.cafe.erp.modules.inventory.material.dto;
 import com.cafe.erp.common.enums.MaterialType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class MaterialUpdateRequest {
     private String name;
@@ -10,7 +12,7 @@ public class MaterialUpdateRequest {
     private Long categoryId;
     private String uomCode;
     private MaterialType materialType;
-    private Double unitCost;
+    private BigDecimal unitCost;
     private Double reorderLevel;
     private Boolean isActive;
 }
